Move user related components to specific package
