+++ src/app/app.module.ts
@@ -8,8 +8,11 @@ import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app.routing.module';
 import { HeaderComponent } from './header/header.component';
 import { DropdownDirective } from './shared/dropdown.directive';
-import { ProfileComponent } from './profile/profile.component';
-import { EditProfileComponent } from './edit-profile/edit-profile.component';
+import { ReviewStationComponent } from './reviewstation/reviewstation.component';
+
+// User related components
+import { ProfileComponent } from './user/profile/profile.component';
+import { EditProfileComponent } from './user/edit-profile/edit-profile.component';
 
 import { HomeComponent } from './home/home.component';
 import { NavbarComponent } from './navbar/navbar.component';
@@ -21,24 +24,25 @@ import { LoginComponent } from './auth/login/login.component';
 import { UpdatePassComponent } from './auth/update-pass/update-pass.component';
 
 // Essay related components
-import {CreateEssayComponent} from './essay/create-essay/create-essay.component';
+import { EssayComponent } from './essay/essay.component';
 import { ReviewEssay } from './essay/review-essay/review-essay.component';
+import {CreateEssayComponent} from './essay/create-essay/create-essay.component';
 import { EssayCardComponent } from './essay/essay-card/essay-card.component';
 import { EditEssayComponent } from './essay/edit-essay/edit-essay.component';
 import { DeleteEssayComponent } from './essay/delete-essay/delete-essay.component';
 
 
 // Services
+import { UserService } from './services/user.service';
 import { AuthGuardService } from './auth-guard.service';
 import { EssayService } from './services/essay.service';
+import { ReviewService } from './services/review.service';
 import { AuthenticationService } from './services/authentication.service';
-import { UserService } from './services/user.service';
 
 // Loading environment
 import { environment } from '../environments/environment';
-import { ReviewService } from './services/review.service';
-import { ReviewStationComponent } from './reviewstation/reviewstation.component';
-import { EssayComponent } from './essay/essay.component';
+
+
 
 @NgModule({
   declarations: [
