+++ src/app/app.routing.module.ts
@@ -5,8 +5,8 @@ import { SignupComponent } from './auth/signup/signup.component';
 import { LoginComponent } from './auth/login/login.component';
 import { UpdatePassComponent } from './auth/update-pass/update-pass.component';
 import { AuthGuardService } from './auth-guard.service';
-import { ProfileComponent } from './profile/profile.component';
-import { EditProfileComponent } from './edit-profile/edit-profile.component';
+import { ProfileComponent } from './user/profile/profile.component';
+import { EditProfileComponent } from './user/edit-profile/edit-profile.component';
 import { HomeComponent } from './home/home.component';
 import { WorkstationComponent } from './workstation/workstation.component';
 import { EssayCardComponent } from './essay/essay-card/essay-card.component';
